pyproject.toml
[project]
name = "systematic-review-data-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.5",
    "crossref-commons>=0.0.7",
    "dspy>=3.0.3",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scholarly>=1.7.11",
]

